Refactor UniversalHandler to split entities and providers
